Fix \date. Fix :SearchReset error
Improved the spacing in the date-insertion shortcut.
Got rid of the ":SearchReset is not an editor command" error when
clearing all search highlighting via "\/", by adding a dummy ":Search"
call first, which causes :SearchReset to be registered. This has a side
effect, though: hitting "\/" removes search highlighting one term at a
time, rather than all at once. I think this might be useful, so I'm
leaving it like that. But for future reference, if I ever want the other
behaviour, then the thing to do will be to remove the ":Search" from the
shortcut, and just call ":Search" once in the vimrc.