Fix json syntax highlightling
I was setting the filetype for *.json files to "json" - but that only
works if you have a json plugin, which I don't (and don't particularly
want at the moment because it seems like overkill for the amount of json
editing that I do).
So, simple fix: json is javascript.